+++ devel/xdhdws/xdhdws.h
@@ -41,7 +41,7 @@ namespace xdhdws {
 	class sProxy
 	{
 	private:
-		Q37_MRMDF( xdhcmn::cProxy, C_, Callback_ );
+		Q37_MRMDF( xdhcmn::cUpstream, C_, Callback_ );
 	public:
 		void reset( bso::bool__ P = true )
 		{
@@ -52,7 +52,7 @@ namespace xdhdws {
 			Callback_ = NULL;
 		}
 		E_CVDTOR( sProxy );
-		void Init( xdhcmn::cProxy *Callback )	// 'Callback' is destroyed by a 'delete' when this object is dsetroyed.
+		void Init( xdhcmn::cUpstream *Callback )	// 'Callback' is destroyed by a 'delete' when this object is destroyed.
 		{
 			reset();